visuals are client side
