+++ Content.Client/Atmos/Piping/Crawling/PipeCrawlingSystem.cs
@@ -1,6 +1,7 @@
+using System.Numerics;
 using Content.Shared.Atmos.Piping.Crawling.Components;
 using Content.Shared.Atmos.Piping.Crawling.Systems;
-using Content.Shared.Movement.Events;
+using Content.Shared.SubFloor;
 using Robust.Client.Graphics;
 using Robust.Shared.Player;
 
@@ -8,9 +9,13 @@ namespace Content.Client.Atmos.Piping.Crawling;
 
 public sealed partial class PipeCrawlingSystem : SharedPipeCrawlingSystem
 {
+    [Dependency] private readonly SharedAppearanceSystem _appearance = default!;
     [Dependency] private readonly IOverlayManager _overlay = default!;
+    [Dependency] private readonly SharedTransformSystem _transform = default!;
 
     private PipeCrawlingOverlay _crawlingOverlay = default!;
+
+    private const float VisualRange = 16;
 
     public override void Initialize()
     {
@@ -19,7 +24,21 @@ public sealed partial class PipeCrawlingSystem : SharedPipeCrawlingSystem
         SubscribeLocalEvent<LocalPlayerAttachedEvent>(OnAttached);
         SubscribeLocalEvent<LocalPlayerDetachedEvent>(OnDetached);
 
+        SubscribeLocalEvent<PipeCrawlingVisualsComponent, ComponentInit>(OnVisualsInit);
+        SubscribeLocalEvent<PipeCrawlingVisualsComponent, ComponentRemove>(OnVisualsRemove);
+
         _crawlingOverlay = new();
+    }
+
+    public override void Update(float frameTime)
+    {
+        base.Update(frameTime);
+
+        var query = EntityQueryEnumerator<PipeCrawlingVisualsComponent>();
+        while (query.MoveNext(out var uid, out var _))
+        {
+            _appearance.SetData(uid, SubFloorVisuals.ScannerRevealed, true);
+        }
     }
 
     private void OnAttached(LocalPlayerAttachedEvent args)
@@ -32,12 +51,22 @@ public sealed partial class PipeCrawlingSystem : SharedPipeCrawlingSystem
         RemoveOverlay(args.Entity);
     }
 
+    private void OnVisualsInit(Entity<PipeCrawlingVisualsComponent> ent, ref ComponentInit args)
+    {
+        _appearance.SetData(ent.Owner, SubFloorVisuals.ScannerRevealed, true);
+    }
+
+    private void OnVisualsRemove(Entity<PipeCrawlingVisualsComponent> ent, ref ComponentRemove args)
+    {
+        _appearance.SetData(ent.Owner, SubFloorVisuals.ScannerRevealed, false);
+    }
+
     protected override void UpdateOverlay(Entity<PipeCrawlingComponent?> ent)
     {
         if (!Resolve(ent.Owner, ref ent.Comp, false))
             return;
 
-        if (!CanDrawOverlay(ent!))
+        if (!IsLocalPlayer(ent!))
             return;
 
         if (!_overlay.HasOverlay<PipeCrawlingOverlay>())
@@ -51,7 +80,7 @@ public sealed partial class PipeCrawlingSystem : SharedPipeCrawlingSystem
         if (!Resolve(ent.Owner, ref ent.Comp, false))
             return;
 
-        if (!CanDrawOverlay(ent!))
+        if (!IsLocalPlayer(ent!))
             return;
 
         if (_crawlingOverlay.Crawler != ent!)
@@ -60,11 +89,83 @@ public sealed partial class PipeCrawlingSystem : SharedPipeCrawlingSystem
         _overlay.RemoveOverlay(_crawlingOverlay);
     }
 
-    private bool CanDrawOverlay(Entity<PipeCrawlingComponent> ent)
+    protected override void UpdateVisuals(Entity<PipeCrawlingComponent> ent)
+    {
+        if (!IsLocalPlayer(ent!))
+            return;
+
+        var toRemove = ent.Comp.PipeNet;
+        ent.Comp.PipeNet.Clear();
+
+        var currentPipe = (ent.Comp.CurrentPipe, PipeQuery.Comp(ent.Comp.CurrentPipe));
+        foreach (var pipe in GetVisiblePipes(currentPipe))
+        {
+            EnsureComp<PipeCrawlingVisualsComponent>(pipe);
+
+            ent.Comp.PipeNet.Add(pipe.Owner);
+            toRemove.Remove(pipe);
+        }
+
+        Dirty(ent);
+
+        UpdateOverlay(ent.AsNullable());
+
+        foreach (var pipe in toRemove)
+            RemComp<PipeCrawlingVisualsComponent>(pipe);
+    }
+
+    protected override void DisableVisuals(Entity<PipeCrawlingComponent> ent)
+    {
+        if (!IsLocalPlayer(ent!))
+            return;
+
+        foreach (var pipe in ent.Comp.PipeNet)
+            RemComp<PipeCrawlingVisualsComponent>(pipe);
+
+        ent.Comp.PipeNet.Clear();
+        Dirty(ent);
+
+        RemoveOverlay(ent.AsNullable());
+    }
+
+    private bool IsLocalPlayer(Entity<PipeCrawlingComponent> ent)
     {
         if (Player.LocalSession?.AttachedEntity != ent.Owner)
             return false;
 
         return true;
+    }
+
+    private IEnumerable<Entity<PipeCrawlingPipeComponent>> GetVisiblePipes(Entity<PipeCrawlingPipeComponent> start)
+    {
+        var startPos = _transform.GetWorldPosition(start);
+
+        HashSet<Entity<PipeCrawlingPipeComponent>> visited = [];
+
+        Queue<Entity<PipeCrawlingPipeComponent>> queue = [];
+        queue.Enqueue(start);
+
+        while (queue.TryDequeue(out var currentPipe))
+        {
+            if (!visited.Add(currentPipe))
+                continue;
+
+            var pos = _transform.GetWorldPosition(currentPipe);
+            var distance = Vector2.Distance(pos, startPos);
+            if (distance > VisualRange)
+                continue;
+
+            yield return currentPipe;
+
+            foreach (var (_, connections) in currentPipe.Comp.ConnectedPipes)
+            {
+                foreach (var (_, netConnection) in connections)
+                {
+                    var connection = GetEntity(netConnection);
+                    if (PipeQuery.TryComp(connection, out var pipe))
+                        queue.Enqueue((connection, pipe));
+                }
+            }
+        }
     }
 }
