+++ Content.Client/Atmos/Piping/Crawling/PipeCrawlingVisualsComponent.cs
@@ -1,0 +1,4 @@
+namespace Content.Client.Atmos.Piping.Crawling;
+
+[RegisterComponent]
+public sealed partial class PipeCrawlingVisualsComponent : Component;
